RF_00.000.14 | Freestyle project > joining all classes
