--- src/test/java/school/redrover/FreeStyleProjectNumber16Test.java
@@ -1,25 +1,0 @@
-package school.redrover;
-
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import school.redrover.model.HomePage;
-import school.redrover.runner.BaseTest;
-
-import java.util.List;
-
-public class FreeStyleProjectNumber16Test extends BaseTest {
-
-    @Test
-    public void testFreeStyleProjectExists() {
-
-        String projectName = "New FreeStyle Project";
-
-        List<String> name = new HomePage(getDriver())
-                .clickNewItem()
-                .setItemName(projectName)
-                .selectFreestyleAndClickOk().clickLogo()
-                .getItemList();
-
-        Assert.assertTrue(name.contains(projectName));
-    }
-}
